+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.4.5-SNAPSHOT] - 03.07.2025.
+## [0.4.5-SNAPSHOT] - 04.07.2025.
+
+### Added
+
+- New event and logic for logging DataTransfer events
 
 ### Changed
 
@@ -10,6 +14,7 @@ All notable changes to this project will be documented in this file.
   background
 - Refactored DataTransferStrategy and implementing classes to return CompletableFuture<Void> for transfer method
 - GeneratePresignURL uses BucketCredentials
+- Renamed DataTranferMockObjectUtil to DataTransferMockObjectUtil
 - Updated tests and Postman collection
 
 ## [0.4.4-SNAPSHOT] - 03.07.2025.